Split persistence invariant tests
