+++ crates/oasis7/src/simulator/tests/mod.rs
@@ -71,6 +71,7 @@ mod module_visual;
 mod monotonicity;
 mod native_resolution_contract;
 mod persist;
+mod persist_journal_invariants;
 mod physics_parameters;
 mod power;
 #[cfg(not(target_arch = "wasm32"))]
